Combine OvalTool+RectTool into ShapeTool
