+++ src/helper/tools/shape-tool.js
@@ -1,0 +1,148 @@
+import paper from '@scratch/paper';
+import {styleShape} from '../style-path';
+import {clearSelection} from '../selection';
+import {getSquareDimensions} from '../math';
+import BoundingBoxTool from '../selection-tools/bounding-box-tool';
+import NudgeTool from '../selection-tools/nudge-tool';
+
+/**
+ * Tool for drawing shapes (e.g. rectangles and ovals).
+ */
+class ShapeTool extends paper.Tool {
+    static get TOLERANCE () {
+        return 2;
+    }
+    /**
+     * @param {function} setSelectedItems Callback to set the set of selected items in the Redux state
+     * @param {function} clearSelectedItems Callback to clear the set of selected items in the Redux state
+     * @param {function} setCursor Callback to set the visible mouse cursor
+     * @param {!function} onUpdateImage A callback to call when the image visibly changes
+     * @param {VectorModes} mode This tool's "mode"
+     */
+    constructor (setSelectedItems, clearSelectedItems, setCursor, onUpdateImage, mode) {
+        super();
+        this.setSelectedItems = setSelectedItems;
+        this.clearSelectedItems = clearSelectedItems;
+        this.onUpdateImage = onUpdateImage;
+        this.boundingBoxTool = new BoundingBoxTool(
+            mode,
+            setSelectedItems,
+            clearSelectedItems,
+            setCursor,
+            onUpdateImage
+        );
+        const nudgeTool = new NudgeTool(this.boundingBoxTool, onUpdateImage);
+
+        // We have to set these functions instead of just declaring them because
+        // paper.js tools hook up the listeners in the setter functions.
+        this.onMouseDown = this.handleMouseDown;
+        this.onMouseMove = this.handleMouseMove;
+        this.onMouseDrag = this.handleMouseDrag;
+        this.onMouseUp = this.handleMouseUp;
+        this.onKeyUp = nudgeTool.onKeyUp;
+        this.onKeyDown = nudgeTool.onKeyDown;
+
+        this.shape = null;
+        this.colorState = null;
+        this.isBoundingBoxMode = null;
+        this.active = false;
+    }
+    getHitOptions () {
+        return {
+            segments: true,
+            stroke: true,
+            curves: true,
+            fill: true,
+            guide: false,
+            match: hitResult =>
+                (hitResult.item.data && (hitResult.item.data.isScaleHandle || hitResult.item.data.isRotHandle)) ||
+                hitResult.item.selected, // Allow hits on bounding box and selected only
+            tolerance: ShapeTool.TOLERANCE / paper.view.zoom
+        };
+    }
+    /**
+     * Should be called if the selection changes to update the bounds of the bounding box.
+     * @param {Array<paper.Item>} selectedItems Array of selected items.
+     */
+    onSelectionChanged (selectedItems) {
+        this.boundingBoxTool.onSelectionChanged(selectedItems);
+    }
+    setColorState (colorState) {
+        this.colorState = colorState;
+    }
+    handleMouseDown (event) {
+        if (event.event.button > 0) return; // only first mouse button
+        this.active = true;
+
+        if (this.boundingBoxTool.onMouseDown(
+            event, false /* clone */, false /* multiselect */, false /* doubleClicked */, this.getHitOptions())) {
+            this.isBoundingBoxMode = true;
+        } else {
+            this.isBoundingBoxMode = false;
+            clearSelection(this.clearSelectedItems);
+            this.shape = new this.shapeConstructor({
+                point: event.downPoint,
+                size: 0
+            });
+            styleShape(this.shape, this.colorState);
+        }
+    }
+    handleMouseDrag (event) {
+        if (event.event.button > 0 || !this.active) return; // only first mouse button
+
+        if (this.isBoundingBoxMode) {
+            this.boundingBoxTool.onMouseDrag(event);
+            return;
+        }
+
+        const downPoint = new paper.Point(event.downPoint.x, event.downPoint.y);
+        const point = new paper.Point(event.point.x, event.point.y);
+        const squareDimensions = getSquareDimensions(event.downPoint, event.point);
+        const realSize = downPoint.subtract(point);
+        if (event.modifiers.shift) {
+            this.shape.size = squareDimensions.size.abs();
+        } else {
+            // Setting a negative size messes up the shape's radius, so keep it positive.
+            this.shape.size = realSize.abs();
+        }
+
+        if (event.modifiers.alt) {
+            this.shape.position = downPoint;
+        } else if (event.modifiers.shift) {
+            this.shape.position = squareDimensions.position;
+        } else {
+            this.shape.position = downPoint.subtract(realSize.multiply(0.5));
+        }
+    }
+    handleMouseUp (event) {
+        if (event.event.button > 0 || !this.active) return; // only first mouse button
+
+        if (this.isBoundingBoxMode) {
+            this.boundingBoxTool.onMouseUp(event);
+            this.isBoundingBoxMode = null;
+            return;
+        }
+
+        if (this.shape) {
+            if (this.shape.area < ShapeTool.TOLERANCE / paper.view.zoom) {
+                // Tiny shape created unintentionally?
+                this.shape.remove();
+                this.shape = null;
+            } else {
+                this.shape.selected = true;
+                this.setSelectedItems();
+                this.onUpdateImage();
+                this.shape = null;
+            }
+        }
+        this.active = false;
+    }
+    handleMouseMove (event) {
+        this.boundingBoxTool.onMouseMove(event, this.getHitOptions());
+    }
+    deactivateTool () {
+        this.boundingBoxTool.deactivateTool();
+    }
+}
+
+export default ShapeTool;
